feat: beter organized the project
