+++ src/js/getLocalLanguage.js
@@ -1,0 +1,5 @@
+function getLocalLanguage() {
+    return localStorage.getItem('language') || 'en';
+}
+
+export default getLocalLanguage;
