+++ src/js/keyboardPressedEvents.js
@@ -1,17 +1,4 @@
-function changeToPressed(button, forceShiftToggle = false) {
-    const shiftLeft = document.querySelector('.ShiftLeft');
-    const shiftRight = document.querySelector('.ShiftRight');
-
-    if (button.classList.contains('ShiftLeft') || button.classList.contains('ShiftLeft') || forceShiftToggle) {
-        shiftLeft.classList.toggle('keyboard__button_activate');
-        shiftRight.classList.toggle('keyboard__button_activate');
-    } else if (button.classList.contains('CapsLock')) {
-        button.classList.toggle('keyboard__button_activate');
-    } else {
-        button.classList.add('active');
-    }
-    return false;
-}
+import changeToPressed from './changeToPressed';
 
 function keyboardPressedEvents(key) {
     const keyCode = `.${key.code}`;
@@ -47,4 +34,4 @@ function keyboardPressedEvents(key) {
 
 
 
-export default { keyboardPressedEvents, changeToPressed };
+export default keyboardPressedEvents;
