+++ src/js/keyboardUnpressedEvents.js
@@ -1,19 +1,4 @@
-function changeToUnpressed(button, forceShiftToggle = false) {
-    const shiftLeft = document.querySelector('.ShiftLeft');
-    const shiftRight = document.querySelector('.ShiftRight');
-
-    if (button.classList.contains('ShiftLeft') || button.classList.contains('ShiftLeft') || forceShiftToggle) {
-        shiftLeft.classList.toggle('keyboard__button_activate');
-        shiftRight.classList.toggle('keyboard__button_activate');
-    } else if (button.classList.contains('CapsLock')) {
-        button.classList.toggle('keyboard__button_activate');
-    } else {
-        button.classList.remove('active');
-        // SetTimeout ?
-    }
-    return false;
-}
-
+import changeToUnpressed from './changeToUnpressed';
 
 function keyboardUnpressedEvents(key) {
     const keyCode = `.${key.code}`;
@@ -30,4 +15,4 @@ function keyboardUnpressedEvents(key) {
     this.pressedButton.delete(key.code);
 }
 
-export default { changeToUnpressed, keyboardUnpressedEvents };
+export default keyboardUnpressedEvents;
