+++ src/scss/base/_buttons.scss
@@ -42,7 +42,7 @@
     .on {
         display: block;
     }
-    .active {
+    &.active {
         background-color: rgba(255, 215, 0, 0.5);
         border: 2px solid rgba(223, 179, 10, 0.5);
         transform: scale(.90);
